#1 #2 #3 - Fixed Multiple Minor Issues
* Small fix for reading temperature data - now working properly.
* Close #1

* Fixed transmission conflicts by adding a small delay. Seems to be
working reliably in lab.
* Close #2

* Implemented a method for storing temperature data.
* Still needs tested (#3)

In addition to the changes from above, I've also done some housekeeping
to make the code readable. Some functionality has been moved out of the
interrupts to make the code more reliable, which has helped somewhat
with conflicting transmissions.

LPM3 seems to be a good way to go. It saves about 3mV of voltage draw
over LPM0, which corresponds to about .9uW. It's not much, but it's
something.

I will be adding test images from lab in the next commit.