Close #3, Close #5. Added RSSI/Identifiers. Finished Temp Sensor
I ultimately decided that the temp sensing is fine as is. Since I'm not
really using it for a directed purpose, fine tuning the storage and
access of the temperature data seemed like a bit of a waste.

Instead, I focused on figuring out how to analyze communication
performance. I decided to enumerate the packets to make identifying a
missing packet really simple in the serial terminal.

I was also browsing the datasheet and found that the module is actually
storing RSSI and that it can be accessed by reading the register and
performing some necessary conversions, so I went ahead and added this
in.

I also made a few changes to the radio settings that I will be testing
on. Since it's kind of a neat scheme, I decided to test using GFSK
first.

A couple of other changes include upping the debounce time, as multiple
sends at once still seem to occur from time to time. I also moved the
print functions and as many delays as possible out of the interrupts, as
this is better practice in general.